requirements.txt
Django==3.0.5
model_bakery==1.1.0
codecov==2.0.22
django-redis-cache==2.1.0
django-redis==4.11.0
redis==3.4.1
djangorestframework==3.11.0
django-registration-redux==2.7
django-registration==3.1
user-agents==2.1
celery==4.4.2
django-celery-results==1.2.1
django-celery-beat==2.0.0
django-influxdb-metrics==1.4.0
